feat: Data init
